Improve markup of dates.
Mark up a date with <time> but without a 'datetime' attribute, if it was
specified in a valid HTML date format. Don't try to mark up a date if it
was specified in any other format, for now, as we don't have a good way
of doing so that preserves the specified precision (day vs. month vs.
year).

Don't mark up 'Present' at all, as marking up with a plain <time>
wrapper was invalid and converting it to today's date seems like a poor
approach.